Shorten via walrus operator
This will do the same thing as before but shorten the code by using the walrus operator. In case you didn't know already, the walrus operator just allows a variable to be set within a larger expression, such as print statements.
PLEASE NOTE: These changes will only work on Python 3.8 or above. If you do not intend on running this code in Python 3.8 or above, please discard this pull request.